Cast the size to uint64_t to make it a known size
Not all platforms have the format specifier to print size_t

Reviewed-on: http://review.couchbase.org/31198
Tested-by: Trond Norbye <trond.norbye@gmail.com>

+++ src/dbdump.c
@@ -145,7 +145,7 @@ static int visit_node(Db *db,
         int *count;
         /* This is a document: */
         printf("%c (%"PRIu64") ", (docinfo->deleted ? 'x' : '*'),
-               docinfo->size);
+               (uint64_t)docinfo->size);
         if (mode == DumpBySequence) {
             printf("#%"PRIu64" ", docinfo->db_seq);
         }
